% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rterm.R
\name{addData}
\alias{addData}
\title{Add data to a "term" object}
\usage{
addData(term, data, formula = NULL, interval = NULL, energyVar = NULL,
  dateStartVar = NULL, dateEndVar = NULL, daysVar = NULL, daily = TRUE)
}
\arguments{
\item{term}{the term object to add data}

\item{data}{the dataset to add}

\item{formula}{an optional specification for energy variable and date variables,
for example kwhd ~ dateStart + dateEnd.}

\item{interval}{an optional specification for interval data. "daily" or "monthly"}

\item{energyVar}{an optional manual specification of the energy variable of the
dataset.}

\item{dateStartVar}{an optional manual specification of the start date variable.}

\item{dateEndVar}{an optional manual specification of the end date variable.}

\item{daysVar}{an optional specification of the days in cycle variable (mainly
for billing data). Note that you only need to specify any two of dateStartVar,
dateEndVar, and daysVar if you do the manual specification.}

\item{daily}{whether or not the energy use is total across the interval or daily.
 Mostly this is daily by the time the analyst gets it, but if dealing with raw
 consumption specify daily = FALSE}
}
\value{
the term object with data added
}
\description{
Adds a dataset & instructions to a term object, returns the updated object
}
\examples{
# Ecotope Mothership data
data(ecotope)
mod <- newTerm("Ecotope Mothership")
mod <- addData(mod, ecotope, kwhd ~ dateStart + dateEnd)
mod <- addData(mod, ecotope, energyVar = "kwhd", dateStartVar = "dateStart", dateEndVar = "dateEnd")
}
\seealso{
\code{\link{newTerm}} \code{\link{addWeather}} \code{\link{addMethod}}
}

