% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rterm.R
\name{addWeather}
\alias{addWeather}
\title{Add weather to a "term" object}
\usage{
addWeather(term, weather = NULL, formula = NULL, stationid = NULL,
  timeVar = NULL, tempVar = NULL, name = NULL)
}
\arguments{
\item{term}{the term object to add data}

\item{weather}{optional dataset if providing your own weather (rather than
using the built-in integration with NOAA).}

\item{formula}{optional specification of time variable and temp variable if
providing your own weather. Example temp ~ date.}

\item{stationid}{optional GHCN ID as found from \code{\link{stationSearch}}.}

\item{timeVar}{optional manual specification of date variable if providing your
own weather data.}

\item{tempVar}{optional manual specification of temp variable if providing your
own weather data.}

\item{name}{optional name for this particular weather. Example name = "Sea-Tac"
 is way better than "SEATTLE-TACOMA INTERNATIONAL AIRPORT"}
}
\value{
the term object w/ weather added
}
\description{
Adds a weather & instructions to a term object, returns the updated object.
If you specified an address in the term construction with \code{\link{newTerm}}
then this function will take care of everything automatically. Note that you
can add multiple weather stations to the "term" object with repeated calls of
this function.
}
\examples{
# Add weather
stationSearch("Seattle, WA")
mod <- addWeather(mod, stationid = "GHCND:USW00024234")
}
\seealso{
\code{\link{newTerm}} \code{\link{addData}} \code{\link{addMethod}}
}

