% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cplm.R
\name{coef.cplm}
\alias{coef.cplm}
\title{Change Point Linear Model Coefficients}
\usage{
\method{coef}{cplm}(x, fit = NULL)
}
\arguments{
\item{x}{a 'cplm' object}

\item{fit}{an optional specification of coefficient type "LS" or "L1". If
  not specified returns the best-guess as to which is more appropriate.
  (The presence of outlying bills trips the standard output to the penalized
  "L1" fit)}
}
\value{
The coefficients from the change point linear model w/ specified
 fitting method.
}
\description{
Change Point Linear Model Coefficients
}
\examples{
data(rfm)
mod <- cplm(eui ~ oat, data = rfm)
coef(mod)
coef(mod, "L1")
coef(mod, "LS")
}

