% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{commercial}
\alias{commercial}
\title{Billed energy use for a Seattle-area commercial building}
\format{A data frame with 36 rows and 3 variables:
\describe{
  \item{Start.Date}{Start date of meter interval}
  \item{End.Date}{End date of meter interval}
  \item{kWh}{Total kWh during the billing cycle}
  ...
}}
\usage{
commercial
}
\description{
A dataset of billed energy use for a commercial building
in the Seattle area
}
\examples{
data(commercial)
head(commercial)
stationSearch("Seattle, WA")
com <- newTerm("Seattle Commercial Building") \%>\%
addData(commercial[-c(1:5), ],
 kWh ~ Start.Date + End.Date,
 daily = FALSE) \%>\%
 addWeather(stationid = "GHCND:USW00024233") \%>\%
 addMethod("Changepoint") \%>\%
 addMethod("degreeday") \%>\%
 evaluate()
com
plot(com, "raw")
plot(com)
plot(com, "resids")
}
\keyword{datasets}

