% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cplm.R
\name{cplm}
\alias{cplm}
\title{Fit a change point linear model}
\usage{
cplm(formula, data, weights, heating = NULL, cooling = NULL, se = TRUE,
  nreps = 200, parametric = NULL, lambda = 0)
}
\arguments{
\item{formula}{a formula as would be used in a linear model}

\item{data}{the dataset to perform the regression with}

\item{weights}{an optional vector of observation weights, if non-NULL will use these for weighted least squares}

\item{heating}{optional to force evaluation of a heating change point}

\item{cooling}{optional to force evaluation of a cooling change point}

\item{se}{estimate standard errors with a bootstrap re-sampling technique}

\item{nreps}{number of bootstrap replicates, defaults to 200}

\item{parametric}{specify true for a parametric bootstrap, FALSE for a
non-parametric bootstrap. Defaults to parametric for < 100 observations,
non-parametric for >= 100 observations}

\item{lambda}{optional override for L1 penalty. Modifies the mean-squared
  error from a full least-squares fit. Larger values correspond to larger
  penalties. A value of 0 corresponds to ordinary least-squares.}
}
\value{
An object of class 'cplm'. Contains the original data.frame as 'dataOrig',
the model formula, the regression data.frame w/ truncated basis vars as 'data',
Least-Squares coefficients as 'LS', L1 penalized coefficients as 'L1', and
optionally a data.frame of 'bootstraps' if se = TRUE.

The following methods have been implemented for the 'cplm' class:
print, coef, predict, plot, resids
}
\description{
Given a formula and data frame, this function will by default
look for heating and cooling with an L1 penalized least-squares
regression, then based on those results fit an ordinary least-
squares regression of the selected model.
}
\examples{
data(rfm)
mod <- cplm(eui ~ oat, data = rfm)
summary(mod)
coef(mod, "LS")
coef(mod, "L1")
plot(mod)
}
\seealso{
\code{\link{plot.cplm}} to plot model output,
  \code{\link{residsPlot}} to plot residuals energy use (net of weather),
  \code{\link{summary.cplm}} to report coefficients + standard errors
    if calculated.
}

