% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cplm.R
\name{cplmx}
\alias{cplmx}
\title{Fit Multiple Change Point Linear Models from a Single Dataset}
\usage{
cplmx(formula, data, id_vars, heating = NULL, cooling = NULL, se = FALSE,
  lambda = 0)
}
\arguments{
\item{formula}{a formula as would be used in a linear model}

\item{data}{the dataset to perform the regression with}

\item{heating}{optional to force evaluation of a heating change point}

\item{cooling}{optional to force evaluation of a cooling change point}

\item{se}{estimate standard errors with a bootstrap re-sampling technique}

\item{lambda}{optional override for L1 penalty. Modifies the mean-squared
  error from a full least-squares fit. Larger values correspond to larger
  penalties. A value of 0 corresponds to ordinary least-squares.}

\item{weights}{an optional vector of observation weights, if non-NULL will use these for weighted least squares}

\item{nreps}{number of bootstrap replicates, defaults to 200}

\item{parametric}{specify true for a parametric bootstrap, FALSE for a
non-parametric bootstrap. Defaults to parametric for < 100 observations,
non-parametric for >= 100 observations}
}
\value{
The coefficients from the change point linear model w/ specified
 fitting method.
}
\description{
This function (change point linear model "extended") wraps \code{\link{cplm}}
for a dataset with multiple buildings.
}
\examples{
data(dhp)
results <- cplmx(kwhd ~ avetemp, data = dhp,
                   id_vars = c("id", "post"))
summary(results)
plot(results, "heatingBase")
}

