% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ecotope}
\alias{ecotope}
\title{Billed energy use for Ecotope's Seattle office}
\format{A data frame with 49 rows and 5 variables:
\describe{
  \item{dateStart}{Start date of meter interval}
  \item{dateEnd}{End date of meter interval}
  \item{days}{Number of days in billing cycle}
  \item{kwhd}{Daily kWh during the billing cycle}
  \item{oat}{Average outdoor temperature}
  ...
}}
\usage{
ecotope
}
\description{
A dataset of billed energy use for the Ecotope Seattle office
from 2006 to 2014
}
\keyword{datasets}

