% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rterm.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a term}
\usage{
evaluate(term)
}
\arguments{
\item{term}{the term object to evaluate.}
}
\value{
the evaluated term object
}
\description{
Evaluates a "term" that has had data, weather, and method(s) added
}
\seealso{
\code{\link{newTerm}} \code{\link{addData}}
\code{\link{addWeather}} \code{\link{addMethod}}
}

