% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rterm.R
\name{extractModel}
\alias{extractModel}
\title{Extract an individual model from a fitted term}
\usage{
extractModel(term, name)
}
\arguments{
\item{term}{the term object to extract from.}

\item{name}{a regex matching the name of the model to extract}
}
\value{
the extracted model
}
\description{
Pulls on individual model fit out of a term object. For example
if you fitted both "degreeday" and "changepoint" you can pull out
just the change point model with extractModel(mod, "cp")
}
\seealso{
\code{\link{newTerm}} \code{\link{addData}}
\code{\link{addWeather}} \code{\link{addMethod}} \code{\link{evaluate}}
}

