% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rterm.R
\name{newTerm}
\alias{newTerm}
\title{Initialize a new "term" - temperature energy regression model}
\usage{
newTerm(name = NULL, address = NULL, sqft = NULL)
}
\arguments{
\item{name}{an optional name for the term. This will go in tables and graphs}

\item{address}{an optional address. If specified then the \code{\link{addWeather}}
function will automatically grab the closest weather station.}

\item{sqft}{an optional building size. If you specify a square feet, it will assume
 that you're dealing in total kWh and will convert that to EUI}
}
\value{
the new term object
}
\description{
Returns a new term object to populate with data, weather, and methods
}
\examples{
# Show available datasets
newTerm("Ecotope Mothership", "4056 9th Ave NE Seattle, WA", 2600)
}
\seealso{
\code{\link{addData}} \code{\link{addWeather}} \code{\link{addMethod}}
}

