% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cplm.R
\name{plot.cplm}
\alias{plot.cplm}
\title{Plot Change Point Linear Model}
\usage{
\method{plot}{cplm}(x, fit = NULL)
}
\arguments{
\item{x}{a 'cplm' object}

\item{fit}{an optional specification of coefficient type "LS" or "L1". If
  not specified returns the best-guess as to which is more appropriate.
  (The presence of outlying bills trips the standard output to the penalized
  "L1" fit.) Additionally specifying "both" here will overlay both fits.}
}
\value{
A ggplot object
}
\description{
Plot Change Point Linear Model
}
\examples{
data(rfm)
mod <- cplm(eui ~ oat, data = rfm)
plot(mod)
plot(mod, "both")
}
\seealso{
\code{\link{residsPlot}} to plot residuals (this is the fun part
  anyway - looking at trends net of weather)
}

