% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weatherHelper.R
\name{plot.stationComp}
\alias{plot.stationComp}
\title{Plot a comparison of GHCN weather stations}
\usage{
plot.stationComp(sc, days = 14, var = "aveTemp", type = "actual",
  xvar = "date")
}
\arguments{
\item{sc}{stationComp object as returned by \code{\link{stationCompare}}}

\item{days}{The number of days for the rolling average. Optional,
defaults to 14.}

\item{var}{The variable to plot. Optional, defaults to "aveTemp". Can
also be "tmin" or "tmax". See \code{\link{read.ghcn}}}

\item{type}{The type of view, can be "actual" or "relative". "actual"
will plot the recorded temperatures, while "relative" will plot the
temperatures with reference to the mean temperature. "relative" tends
to be more useful in visualizing the differences between stations.
This argument is optional and defaults to "relative".}

\item{xvar}{What variable to plot against. Defaults to "date" but can
  also be "doy" - day of year. "date" gives a long rolling line, while
  "doy" gives multiple lines on top of each other per station.}
}
\value{
a ggplot object of the graphic
}
\description{
Plot a comparison of weather stations returned by
\code{\link{stationCompare}}
}
\examples{
# Compare 5 closest weather stations to Ecotope's Seattle office
stations <- stationSearch("Seattle")
comp <- stationCompare(stations, "2014-01-01", "2014-12-31")
plot(comp)
plot(comp, "days" = 60)
plot(comp, var = "tmin", type = "actual")
}
\seealso{
\code{\link{stationSearch}} \code{\link{stationComp}}
 \code{\link{plot.stationComp}}  \code{\link{read.ghcn}}
}

