% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/models.R
\name{projection}
\alias{projection}
\title{Project a temperature-energy dependence onto archival weather data}
\usage{
projection(term, nYears = 20)
}
\arguments{
\item{nYears}{optional number of years to project onto.}

\item{mod}{An individual model object that has been extracted with \code{\link{extractModel}}}
}
\value{
a projection object that can be printed or plotted
}
\description{
Returns a projection object that can be printed or plotted to deliver a long-
term average prediction of energy use. This function currently only works with
the NOAA API for Daily GHCN data. See vignette("weather-helper") for details on
setting up this functionality
}
\examples{
# Assuming a term has been fit called 'mod' with a NOAA webservices weather
p <- projection(mod, nYears = 20)
p
plot(p)
}

