% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weatherHelper.R
\name{read.ghcn}
\alias{read.ghcn}
\title{Retrive GHCN Weather Data}
\usage{
read.ghcn(stationid, startdate, enddate)
}
\arguments{
\item{stationid}{the identifier for the weather station}

\item{startdate}{the first date of data requested}

\item{enddate}{the last date of data requested}
}
\value{
a data frame with variables "date", "TMIN", "TMAX", and
  "aveTemp". Temperatures in Fahrenheit
}
\description{
Given a station id and start and end dates, retrieves daily
temperature data from NOAA's API. This wraps the function
read.one.ghcn, which reads data one year at a time (that's the
maximum you're allowed to get in a single query), which itself
gets raw data with the \code{\link{read.noaa}} function.
}
\details{
Two things to note: 1) you need a key from NOAA to access the data.
This is free and you can generate one here: \url{https://www.ncdc.noaa.gov/cdo-web/token}.
All weather functions in rterm look for an object in the global environment
called "noaa_key". You can deal with this by taking the following line:

noaa_key <- "your_key_here"

and placing it either before the library(rterm) statement, or, preferably, in
a .Rprofile. \url{http://www.statmethods.net/interface/customizing.html}

2) You need to find a weather station to use! Probably the best way
to do this for now is to use the NOAA web search at
\url{http://www.ncdc.noaa.gov/cdo-web/search}, where you search for
"Daily Summaries" and "Stations" with the appropriate date range and a
keyword. This will bring up a map where you can click on icons for weather
stations and it will report an "ID" that looks something like
"GHCND:US1WAKG0179"
}
\examples{
weather <- read.ghcn("GHCND:USW00024233", "2010-01-01", "2014-12-31")
}
\seealso{
\code{\link{read.noaa}} \code{\link{stationSearch}}
}

