% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{rural}
\alias{rural}
\title{Billed energy use for a rural Washington State home}
\format{A data frame with 49 rows and 5 variables:
\describe{
  \item{start}{Start date of meter interval}
  \item{end}{End date of meter interval}
  \item{days}{Number of days in billing cycle}
  \item{kwh}{Total kWh during the billing cycle}
  \item{kwh.day}{Daily kWh during the billing cycle}
  ...
}}
\usage{
rural
}
\description{
A dataset of billed energy use for a home in rural Washington State
}
\keyword{datasets}

