% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weatherHelper.R
\name{stationCompare}
\alias{stationCompare}
\title{Compare GHCN Weather Stations}
\usage{
stationCompare(st, startdate, enddate)
}
\arguments{
\item{st}{Data frame of stations returned by \code{\link{stationSearch}}}

\item{startdate}{the start date of the interval to compare}

\item{enddate}{the end date of the interval to compare

The start and end date arguments should either be POSIX date objects,
or coercible to with as.Date. So for example "2013-01-01" is appropriate.}
}
\value{
An object of S3 class stationComp. This is a list with two entries
1) data - the weather data for the requested stations over the requested interval
and 2) stations - the station info data frame.
}
\description{
Compare weather stations returned by \code{\link{stationSearch}}
}
\examples{
# Compare 5 closest weather stations to Ecotope's Seattle office
stations <- stationSearch(lat = 47.6569326, lon = -122.3184546)
comp <- stationCompare(stations, "2014-01-01", "2014-12-31")
summary(comp)
plot(comp)
}
\seealso{
\code{\link{summary.stationComp}} \code{\link{plot.stationComp}}
 \code{\link{read.ghcn}}
}

