% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weatherHelper.R
\name{stationTrend}
\alias{stationTrend}
\title{Look at a time trend within a station, for example the past 5 years}
\usage{
stationTrend(weather, var = "aveTemp", days = 60, type = "actual")
}
\arguments{
\item{weather}{a weather dataset returned by read.ghcn}

\item{var}{The variable to plot. Optional, defaults to "aveTemp". Can
also be "tmin" or "tmax". See \code{\link{read.ghcn}}}

\item{days}{The number of days for the rolling average. Optional,
defaults to 14.}

\item{type}{The type of view, can be "actual" or "relative". "actual"
  will plot the recorded temperatures, while "relative" will plot the
  temperatures with reference to the mean temperature. "relative" tends
  to be more useful in visualizing the differences between stations.
  This argument is optional and defaults to "relative".}
}
\value{
a list containing "data" the dataset w/ additional smoothed
 variable(s) and "plot", the ggplot object
}
\description{
Plot a comparison of weather stations returned by
\code{\link{stationCompare}}
}
\examples{
# Look at the last 5 years from Ted Stevens Intl Airport in Alaska
anc <- read.ghcn("GHCND:USW00026451", "2010-01-01", "2015-04-15")
ancTrend <- stationTrend(anc)
ancTrend$plot
}
\seealso{
\code{\link{stationSearch}} \code{\link{stationComp}}
 \code{\link{plot.stationComp}}  \code{\link{read.ghcn}}
}

