% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weatherHelper.R
\name{summary.stationComp}
\alias{summary.stationComp}
\title{Summarize a comparison of GHCN weather stations}
\usage{
summary.stationComp(sc, days = 14)
}
\arguments{
\item{sc}{stationComp object as returned by \code{\link{stationCompare}}}
}
\value{
a data frame summarizing the weather stations. This includes
 the station name and id, as well as some summary stats to hopefully
 help you choose. "dataFrac" refers to the fraction of non-missing data
 in the observed interval ("datacoverage" from stationSearch refers to
 the entire history of the weather station). "relativeTemp" refers to
 the average temperature of that station, with respect to the average
 temperature of all stations in the comparison. If you searched by
 latitude/longitude, you also get "milesDistant", which is miles from your
 search coordinates to the weather station.
}
\description{
Summarize a comparison of weather stations returned by
\code{\link{stationCompare}}
}
\examples{
# Compare 5 closest weather stations to Ecotope's Seattle office
stations <- stationSearch(lat = 47.6569326, lon = -122.3184546)
comp <- stationCompare(stations, "2014-01-01", "2014-12-31")
summary(comp)
}
\seealso{
\code{\link{stationSearch}} \code{\link{stationComp}}
 \code{\link{plot.stationComp}}  \code{\link{read.ghcn}}
}

