% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cplm.R
\name{residsPlot}
\alias{residsPlot}
\title{Plot Residuals from a Change Point Linear Model}
\usage{
residsPlot(x, ...)
}
\arguments{
\item{x}{a 'cplm' object}

\item{var}{variable name to plot against in the dataset used to
  fit the model.}
}
\value{
A ggplot object
}
\description{
Note that this is a generic, currently with a method for
S3 class 'cplm'. The vision is that once variable base degree day
& Bayesian methodologies are developed there will also be methods
to plot the residuals from those models.
}
\examples{
data(ecotope)
mod <- cplm(kwhd ~ oat, data = ecotope)
residsPlot(mod, "dateEnd")
}
\seealso{
\code{\link{plot.cplm}}
}

